nit: Fixed some minor structure issues to make accessors more consistent.
