+++ emily/handler/src/bin/emily-server.rs
@@ -9,7 +9,7 @@ use emily_handler::logging;
 
 #[tokio::main]
 async fn main() {
-    logging::setup_logging("info,emily-handler=debug", true);
+    logging::setup_logging("info,emily-handler=debug", false);
 
     // TODO(389 + 358): Handle config pickup in a way that will only fail for the relevant call.
     let emily_context: EmilyContext = EmilyContext::local_test_instance()
